% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stack_and_formula.R
\name{build_stack_and_formula}
\alias{build_stack_and_formula}
\title{Build data stack and formula for AR model}
\usage{
build_stack_and_formula(Data, x_col, y_col, time_col, response_col, order,
  covariate_cols, ValidationData = NULL, prediction_steps = 0, mesh2d,
  mesh1d)
}
\arguments{
\item{Data}{dataframe containing all variables and covariates and time and location of each observation}

\item{x_col}{name of column containing position on x access (longitude) of each location}

\item{y_col}{name of column containing position on y access (latitude) of each location}

\item{time_col}{name of column containing time point of each observation}

\item{response_col}{name of column containing variable of interest}

\item{order}{order of autoregressive model}

\item{covariate_cols}{vector containing name of all columns containing covariates}

\item{ValidationData}{dataframe containing the data for validating the predictions}

\item{prediction_steps}{Number of steps ahead to be predicted, validation data must also be provided}

\item{mesh2d}{2d mesh}

\item{mesh1d}{1d mesh}
}
\value{
list containing the data stack and the formula
}
\description{
Build data stack and formula for AR model
}
