% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_validation_set.R
\name{create_validation_set}
\alias{create_validation_set}
\title{Splits data validation data}
\usage{
create_validation_set(data, time_col, location_col, response_col,
  covariate_cols, prediction_steps, covariate_change = "Constant")
}
\arguments{
\item{time_col}{name of column containing time point of each observation}

\item{location_col}{name of column containing location of each observation}

\item{response_col}{name of column containing response}

\item{covariate_cols}{name of column containing the value for ab consumption}

\item{prediction_steps}{number of time steps ahead to predict (assumes intervals are constant)}

\item{covariate_change}{Default is Constant. Will assume that ab consumption remains same as last observed time point}

\item{Data}{dataframe containing all variables and covariates and time and location of each observation}
}
\value{
A list containing training data and the validation data
}
\description{
Splits data validation data
}
