% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_real_v_simulated.R
\name{plot_real_v_simulated}
\alias{plot_real_v_simulated}
\title{PLot real versus model simulated data}
\usage{
plot_real_v_simulated(result, stack.est, Data, response_col,
  filename = NULL, return = FALSE)
}
\arguments{
\item{result}{fitted inla model}

\item{Data}{data model was fitted to}

\item{response_col}{name of column in data containing response data}

\item{filename}{file for plot to be saved to, default FALSE}

\item{return}{default if FALSE, if TRUE plot will be reurned}

\item{stack}{data stack}
}
\description{
PLot real versus model simulated data
}
