% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run INLA model}
\usage{
run_model(stack, formula, spde, likelihood = "gaussian")
}
\arguments{
\item{stack}{Data stack}

\item{formula}{ar formula}

\item{likelihood}{default gaussian}
}
\value{
result of inla model
}
\description{
Run INLA model
}
