% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_fit.R
\name{validate_model_fit}
\alias{validate_model_fit}
\title{Computes multiple statistics describing the model fit}
\usage{
validate_model_fit(ValidationData, response_col, stack.est)
}
\arguments{
\item{ValidationData}{dataframe containing the validation data}

\item{response_col}{names of response column}

\item{stack.est}{stack.est object}
}
\description{
Computes multiple statistics describing the model fit
}
