% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_of_fits.R
\name{write_summary_of_fits}
\alias{write_summary_of_fits}
\title{Write summary of multiple model fits to a dataframe}
\usage{
write_summary_of_fits(all_data_files = NULL, all_model_fits = NULL)
}
\arguments{
\item{all_data_files}{list of file names all containing one fited inla model object}

\item{all_model_fits}{list of fitted inla model objects}
}
\value{
dataframe containing aggregated fit summaries
}
\description{
Write summary of multiple model fits to a dataframe
}
